+++ core/src/main/java/io/bitsquare/payment/SepaAccountContractData.java
@@ -40,7 +40,7 @@ public final class SepaAccountContractData extends CountryBasedPaymentAccountCon
     // Dont use a set here as we need a deterministic ordering, otherwise the contract hash does not match
     private final ArrayList<String> acceptedCountryCodes;
 
-    public SepaAccountContractData(String paymentMethod, String id, int maxTradePeriod) {
+    public SepaAccountContractData(String paymentMethod, String id, long maxTradePeriod) {
         super(paymentMethod, id, maxTradePeriod);
         Set<String> acceptedCountryCodesAsSet = CountryUtil.getAllSepaCountries().stream().map(e -> e.code).collect(Collectors.toSet());
         acceptedCountryCodes = new ArrayList<>(acceptedCountryCodesAsSet);
